+++ packages/connectors/hull-browser/src/destinations/index.js
@@ -1,7 +1,0 @@
-// import ga from "./google-analytics";
-
-export default function handleDestinations(_emitter) {
-  // emitter.on("user.update", update => {
-  //   ga(update);
-  // });
-}