+++ packages/connectors/hull-outreach/server/transforms-to-hull.js
@@ -1,7 +1,15 @@
 /* @flow */
 import type { ServiceTransforms } from "hull-connector-framework/src/purplefusion/types";
 
-const { doesNotContain, isEqual, doesContain, isNotEqual, isServiceAttribute, mappingExists } = require("hull-connector-framework/src/purplefusion/conditionals");
+const {
+  doesNotContain,
+  isEqual,
+  mappingExists,
+  notNull,
+  isNull,
+  not,
+  resolveIndexOf
+} = require("hull-connector-framework/src/purplefusion/conditionals");
 
 
 const {
@@ -407,6 +415,28 @@ const transformsToHull: ServiceTransforms =
       strategy: "MixedTransforms",
       transforms: [
         {
+          strategy: "AtomicReaction",
+          target: { component: "input", name: "eventInput" },
+          validation:
+            {
+              error: "BreakProcess",
+              message: "Event has never been seen before by the connector, please report issue to your Hull Support representative",
+              condition:
+                doesNotContain(require("./events.json"), "eventInput.attributes.name")
+            }
+        },
+        {
+          strategy: "AtomicReaction",
+          target: { component: "input", name: "eventInput" },
+          validation:
+            {
+              error: "BreakToLoop",
+              message: "Event has not been whitelisted by the connector settings, please see the \"Events To Fetch\" in the settings to add this event type",
+              condition:
+                not(resolveIndexOf("connector.private_settings.events_to_fetch", "eventInput.attributes.name"))
+            }
+        },
+        {
           strategy: "Jsonata",
           direction: "incoming",
           transforms: [
@@ -444,21 +474,29 @@ const transformsToHull: ServiceTransforms =
           target: { component: "input", name: "eventInput" },
           then: [
             {
-              validation: {
-                error: "BreakProcess",
-                message: "Event has never been seen before by the connector, please report issue to your Hull Support representative",
-                condition:
-                  doesNotContain(require("./events.json"), "eventInput.hull_events[0].eventName")
-              }
+              //target: { component: "input", select: "hullEvents[0]"},
+              condition: notNull("eventInput.hull_events[0].properties.email_id"),
+              operateOn: { component: "input", name: "mailingId", select: "hull_events[0].properties.email_id" },
+              then: [
+                {
+                  operateOn: { component: "glue", route: "getMailingDetails", name: "enrichedEmail" },
+                  // writeTo: { path: "hull_events[0].properties.email_subject", format: "${enrichedEmail.email_subject}" },
+                  then: [
+                    {
+                      writeTo: { path: "hull_events[0].properties.email_subject", format: "${enrichedEmail.email_subject}" }
+                    },
+                    {
+                      writeTo: { path: "hull_events[0].properties.sequence_id", format: "${enrichedEmail.sequence_id}" },
+                    },
+                    {
+                      // condition: notNull("${enrichedEmail.sequence_id}"),
+                      operateOn: { component: "glue", route: "getSequences", select: "${enrichedEmail.sequence_id}" },
+                      writeTo: { path: "hull_events[0].properties.sequence_name" }
+                    },
+                  ]
+                },
+              ]
             },
-            // {
-            //   validation: {
-            //     error: "BreakToLoop",
-            //     message: "Event has not been whitelisted by the connector settings, please see the \"Events To Fetch\" in the settings to add this event type",
-            //     condition:
-            //       mappingExists("events_to_fetch", "eventInput.hull_events[0].eventName")
-            //   }
-            // },
             {
               operateOn: {
                 component: "static",
@@ -474,9 +512,13 @@ const transformsToHull: ServiceTransforms =
                 select: "${eventInput.hull_events[0].eventName}",
                 name: "eventName"
               },
-              // validation: { error: "BreakProcess", condition: [
-              //     isUndefinedOrNull("${eventName}"),
-              //   ]},
+              then: [
+                {
+                  validation: { error: "BreakToLoop", condition: [
+                      isNull("eventName"),
+                    ]},
+                }
+              ],
               writeTo: {
                 path: "hull_events[0].eventName"
               }