+++ packages/connectors/hull-zapier/zapier-app/src/test/input-fields/account-input-fields.test.js
@@ -9,7 +9,7 @@ process.env.TOKEN = "YOUR_TOKEN";
 const App = require('../../index');
 const appTester = zapier.createAppTester(App);
 
-describe('Input Fields Test', () => {
+describe('Account Input Fields Test', () => {
   zapier.tools.env.inject();
   const connector_url = process.env.CONNECTOR_URL;
 
@@ -172,7 +172,7 @@ describe('Input Fields Test', () => {
     // Account Schema
     scope.post("/schema?token=YOUR_TOKEN")
       .reply(200, [
-        { "label": "domain", "value": "domain" }
+        { "name": "account.domain" }
       ]);
 
     const results = await appTester(