+++ packages/connectors/hull-zapier/zapier-app/src/test/input-fields/user-input-fields.test.js
@@ -9,7 +9,7 @@ process.env.TOKEN = "YOUR_TOKEN";
 const App = require('../../index');
 const appTester = zapier.createAppTester(App);
 
-describe('Input Fields Test', () => {
+describe('User Input Fields Test', () => {
   zapier.tools.env.inject();
   const connector_url = process.env.CONNECTOR_URL;
 
@@ -181,17 +181,14 @@ describe('Input Fields Test', () => {
     // User Schema
     scope.post("/schema?token=YOUR_TOKEN")
       .reply(200, [
-        // TODO remove "account." traits from user
-        { "label": "account.domain", "value": "account.domain" },
-        { "label": "account.salesforce/industry", "value": "account.salesforce/industry" },
-        { "label": "email", "value": "email" },
-        { "label": "salesforce_contact/department", "value": "salesforce_contact/department" }
+        { "name": "user.email" },
+        { "name": "user.salesforce_contact/department" }
       ]);
 
     // Account Schema
     scope.post("/schema?token=YOUR_TOKEN")
       .reply(200, [
-        { "label": "domain", "value": "domain" }
+        { "name": "account.domain" }
       ]);
 
     const results = await appTester(
@@ -230,14 +227,6 @@ describe('Input Fields Test', () => {
           "label": "User Attributes",
           "list": true,
           "choices": [
-            {
-              "label": "account.domain",
-              "value": "account.domain"
-            },
-            {
-              "label": "account.salesforce/industry",
-              "value": "account.salesforce/industry"
-            },
             {
               "label": "email",
               "value": "email"
@@ -312,9 +301,9 @@ describe('Input Fields Test', () => {
     // User Events
     scope.post("/schema?token=YOUR_TOKEN")
       .reply(200, [
-        {"label": "Email Opened", "value": "Email Opened"},
-        {"label": "Email Sent", "value": "Email Sent"},
-        {"label": "Email Dropped", "value": "Email Dropped"}
+        {"name": "user_event.Email Opened"},
+        {"name": "user_event.Email Sent"},
+        {"name": "user_event.Email Dropped"}
       ]);
 
     const results = await appTester(