+++ packages/hull-connector-framework/src/purplefusion/dispatcher.js
@@ -332,7 +332,7 @@ class HullDispatcher {
 
             // if we get stop, that means that there was an "if" condition
             // and it did not validate, which meant we return stop
-            if (elifResult.status !== "stop") {
+            if (elifResult.hullDispatcherStatus !== "stop") {
               return {};
             }
           }
@@ -344,7 +344,7 @@ class HullDispatcher {
           return await this.resolve(context, instructionOptions.eldo, serviceData);
         }
 
-        return { status: "stop" };
+        return { hullDispatcherStatus: "stop" };
 
       } else if (instructionName === "filter") {
 
@@ -404,7 +404,13 @@ class HullDispatcher {
             if (setKey) {
               shallowContextClone.set(setKey, resolvedParams[key]);
             }
-            return this.resolve(shallowContextClone, instructionOptions.instructions, serviceData);
+            return this.resolve(shallowContextClone, instructionOptions.instructions, serviceData)
+              .catch(err => {
+                if (err.code === "BreakToLoop") {
+                  return Promise.resolve();
+                }
+                return Promise.reject(err);
+              });
           }));
         }
 
@@ -454,7 +460,7 @@ class HullDispatcher {
               finalInstruction = instructionResults;
             }
           } catch(err) {
-            if (typeof err !== "BreakToLoop") {
+            if (err.code !== "BreakToLoop") {
               throw err;
             }
           }